+++ src/Enums/paths.js
@@ -1,0 +1,14 @@
+/* -------------------------------------
+ * This object contains all the paths
+ * to pages required by this project.
+ * -------------------------------------
+ * Key Terms:
+ * -> STT - Semester Timetable page
+ * -> SEM - Information about semester
+ */
+const paths = {
+  STT: 'Semesters/semester_info/SEMESTER_TIMETABLE.php',
+  SEM: 'Semesters/ajax.php',
+};
+
+module.exports = paths;
